pass the request object and add the base method
the Response object can accept a request object, that will act as a property.
A new method is added: `base`. This will returns the value of the url for the
response. The value can be set in headers (Content-Base or Content-Location).
If none of the headers are set, then it will use the value for `url` in the
request object.